Apply missing overlay shape and refactor demo
